docs: display __init__() method docstrings
Display the class' __init__ methods' docstrings in the online
documentation. This helps users to understand the class APIs.

Remove the docstring from the _XYZPacket parent class, because Sphinx
was rendering that for every child class (AccelerometerPacket,
GyroPacket, MagnetometerPacket, QuaternionPacket). Apparently the
"autodoc_inherit_docstrings = False" setting does not work in that case.